# tutorflow-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tutorflow_tutorflow-server_internal_domain.Assignment:
    properties:
      allow_late_submission:
        type: boolean
      allowed_file_types:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      instructions:
        type: string
      late_penalty_percent:
        type: number
      lesson_id:
        type: string
      max_file_size:
        description: 10MB
        type: integer
      max_score:
        type: number
      submissions:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Submission'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      parent:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Category'
      parent_id:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      subcategories:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Category'
        type: array
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.Certificate:
    properties:
      certificate_number:
        type: string
      enrollment:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      issued_at:
        type: string
      pdf_url:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.ContentAccess:
    enum:
    - free
    - enrolled
    - premium
    type: string
    x-enum-varnames:
    - ContentAccessFree
    - ContentAccessEnrolled
    - ContentAccessPremium
  github_com_tutorflow_tutorflow-server_internal_domain.Conversation:
    properties:
      course:
        allOf:
        - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
        description: Relationships
      course_id:
        description: 'Optional: context for conversation'
        type: string
      created_at:
        type: string
      id:
        type: string
      last_message:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Message'
      last_message_at:
        type: string
      messages:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Message'
        type: array
      participant1_id:
        type: string
      participant2_id:
        type: string
      updated_at:
        type: string
      user1:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
      user2:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_domain.Course:
    properties:
      categories:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Category'
        type: array
      created_at:
        type: string
      description:
        type: string
      discount_price:
        type: number
      duration_hours:
        type: integer
      enrollments:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Enrollment'
        type: array
      id:
        type: string
      instructor:
        allOf:
        - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
        description: Relationships
      instructor_id:
        type: string
      is_featured:
        type: boolean
      language:
        type: string
      level:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.CourseLevel'
      modules:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Module'
        type: array
      preview_video_url:
        type: string
      price:
        type: number
      published_at:
        type: string
      rating:
        type: number
      requirements:
        items:
          type: string
        type: array
      reviews:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.CourseReview'
        type: array
      short_description:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.CourseStatus'
      thumbnail_url:
        type: string
      title:
        type: string
      total_lessons:
        type: integer
      total_reviews:
        type: integer
      total_students:
        type: integer
      updated_at:
        type: string
      what_you_learn:
        items:
          type: string
        type: array
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.CourseLevel:
    enum:
    - beginner
    - intermediate
    - advanced
    type: string
    x-enum-varnames:
    - CourseLevelBeginner
    - CourseLevelIntermediate
    - CourseLevelAdvanced
...
  github_com_tutorflow_tutorflow-server_internal_domain.CourseReview:
    properties:
      content:
        type: string
      course:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
      course_id:
        type: string
      created_at:
        type: string
      helpful_count:
        type: integer
      id:
        type: string
      instructor_reply:
        type: string
      instructor_reply_at:
        type: string
      is_featured:
        type: boolean
      is_verified_purchase:
        type: boolean
      rating:
        description: 1.0 - 5.0
        type: number
      status:
        type: string
      title:
        type: string
      unhelpful_count:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
      user_id:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.CourseStatus:
    enum:
    - draft
    - published
    - archived
    type: string
    x-enum-varnames:
    - CourseStatusDraft
    - CourseStatusPublished
    - CourseStatusArchived
...
  github_com_tutorflow_tutorflow-server_internal_domain.Enrollment:
    properties:
      certificate:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Certificate'
      completed_at:
        type: string
      course:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
      course_id:
        type: string
      enrolled_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_accessed_at:
        type: string
      order_id:
        description: Link to purchase
        type: string
      progress:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.LessonProgress'
        type: array
      progress_percent:
        type: number
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.EnrollmentStatus'
      user:
        allOf:
        - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
        description: Relationships
      user_id:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.EnrollmentStatus:
    enum:
    - pending
    - active
    - completed
    - cancelled
    - expired
    type: string
    x-enum-varnames:
    - EnrollmentStatusPending
    - EnrollmentStatusActive
    - EnrollmentStatusCompleted
    - EnrollmentStatusCancelled
    - EnrollmentStatusExpired
...
  github_com_tutorflow_tutorflow-server_internal_domain.Lesson:
    properties:
      access_type:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.ContentAccess'
      assignment:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Assignment'
      attachments:
        type: string
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_preview:
        type: boolean
      is_published:
        type: boolean
      lesson_type:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.LessonType'
      module:
        allOf:
        - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Module'
        description: Relationships
      module_id:
        type: string
      quiz:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Quiz'
      sort_order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      video_assets:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.VideoAsset'
        type: array
      video_duration:
        description: seconds
        type: integer
      video_url:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.LessonProgress:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      enrollment_id:
        type: string
      id:
        type: string
      is_completed:
        type: boolean
      lesson:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Lesson'
      lesson_id:
        type: string
      time_spent:
        description: seconds
        type: integer
      updated_at:
        type: string
      video_position:
        description: seconds
        type: integer
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.LessonType:
    enum:
    - video
    - text
    - quiz
    - assignment
    - resource
    type: string
    x-enum-varnames:
    - LessonTypeVideo
    - LessonTypeText
    - LessonTypeQuiz
    - LessonTypeAssignment
    - LessonTypeResource
  github_com_tutorflow_tutorflow-server_internal_domain.Message:
    properties:
      attachment_url:
        type: string
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      read_at:
        type: string
      sender:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
      sender_id:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.Module:
    properties:
      course:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
      course_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_published:
        type: boolean
      lessons:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Lesson'
        type: array
      sort_order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_domain.QuestionType:
    enum:
    - single_choice
    - multiple_choice
    - true_false
    - short_answer
    - essay
    type: string
    x-enum-varnames:
    - QuestionTypeSingleChoice
    - QuestionTypeMultipleChoice
    - QuestionTypeTrueFalse
    - QuestionTypeShortAnswer
    - QuestionTypeEssay
  github_com_tutorflow_tutorflow-server_internal_domain.Quiz:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_published:
        type: boolean
      lesson_id:
        type: string
      max_attempts:
        type: integer
      passing_score:
        type: number
      questions:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.QuizQuestion'
        type: array
      show_correct_answers:
        type: boolean
      shuffle_questions:
        type: boolean
      time_limit:
        description: minutes
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_domain.QuizOption:
    properties:
      id:
        type: string
      is_correct:
        type: boolean
      option_text:
        type: string
      question_id:
        type: string
      sort_order:
        type: integer
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.QuizQuestion:
    properties:
      created_at:
        type: string
      explanation:
        type: string
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.QuizOption'
        type: array
      points:
        type: number
      question_text:
        type: string
      question_type:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.QuestionType'
      quiz_id:
        type: string
      sort_order:
        type: integer
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_domain.Submission:
    properties:
      assignment:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Assignment'
      assignment_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      feedback:
        type: string
      file_name:
        type: string
      file_url:
        type: string
      graded_at:
        type: string
      graded_by:
        type: string
      grader:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
      id:
        type: string
      score:
        type: number
      status:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.SubmissionStatus'
      submitted_at:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
      user_id:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.SubmissionStatus:
    enum:
    - pending
    - submitted
    - graded
    - returned
    type: string
    x-enum-varnames:
    - SubmissionStatusPending
    - SubmissionStatusSubmitted
    - SubmissionStatusGraded
    - SubmissionStatusReturned
  github_com_tutorflow_tutorflow-server_internal_domain.TutorProfile:
    properties:
      created_at:
        type: string
      hourly_rate:
        type: number
      id:
        type: string
      payout_email:
        type: string
      qualifications:
        items:
          type: string
        type: array
      rating:
        type: number
      specializations:
        items:
          type: string
        type: array
      total_courses:
        type: integer
      total_earnings:
        type: number
      total_reviews:
        type: integer
      total_students:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.User'
      user_id:
        type: string
      years_of_experience:
        type: integer
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      courses:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
        type: array
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      enrollments:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Enrollment'
        type: array
      first_name:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.UserRole'
      status:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.UserStatus'
      tutor_profile:
        allOf:
        - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.TutorProfile'
        description: Relationships
      updated_at:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_domain.UserRole:
    enum:
    - admin
    - manager
    - tutor
    - student
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
    - RoleTutor
    - RoleStudent
  github_com_tutorflow_tutorflow-server_internal_domain.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    - pending
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusSuspended
    - StatusPending
  github_com_tutorflow_tutorflow-server_internal_domain.VideoAsset:
    properties:
      bitrate:
        type: integer
      created_at:
        type: string
      drm_policy_id:
        type: string
      duration:
        description: seconds
        type: integer
      encryption_key_id:
        type: string
      file_size:
        type: integer
      hls_playlist_url:
        type: string
      id:
        type: string
      lesson_id:
        type: string
      original_filename:
        type: string
      resolution:
        type: string
      status:
        type: string
      storage_path:
        type: string
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.ValidationError'
        type: array
      message:
        type: string
    type: object
  github_com_tutorflow_tutorflow-server_internal_pkg_response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_tutorflow_tutorflow-server_internal_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.ErrorInfo'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_tutorflow_tutorflow-server_internal_pkg_response.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_usecase_message.SendMessageInput:
    properties:
      attachment_url:
        type: string
      content:
        maxLength: 5000
        minLength: 1
        type: string
      conversation_id:
        type: string
      recipient_id:
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tutorflow.com
    name: API Support
    url: http://www.tutorflow.com/support
  description: This is the backend API for TutorFlow Learning Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TutorFlow LMS API
  version: "1.0"
paths:
...
  /messages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Message data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_message.SendMessageInput'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Message'
              type: object
      security:
      - BearerAuth: []
      summary: Send a message
      tags:
      - Messages
  /messages/{id}:
    delete:
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a message
      tags:
      - Messages
  /messages/{id}/read:
    post:
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Mark a message as read
      tags:
      - Messages
  /messages/conversations:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get my conversations
      tags:
      - Messages
  /messages/conversations/{id}:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Conversation'
              type: object
      security:
      - BearerAuth: []
      summary: Get a conversation
      tags:
      - Messages
  /messages/conversations/{id}/messages:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get messages in a conversation
      tags:
      - Messages
  /messages/conversations/{id}/read:
    post:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Mark all messages in a conversation as read
      tags:
      - Messages
  /messages/conversations/{userId}:
    post:
      parameters:
      - description: User ID to message
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Conversation'
              type: object
      security:
      - BearerAuth: []
      summary: Start a conversation with a user
      tags:
      - Messages
  /messages/unread-count:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get total unread message count
      tags:
      - Messages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"